requirements.txt
google-generativeai
python-dotenv
tavily-python
